# docs/Api-Documentation.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "E-Commerce Ordering & Payment API",
        "description": "Backend for managing users, products, orders, and payments with support for multiple payment providers (Stripe, bKash).",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/users/me": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Read Own Profile",
                "operationId": "read_own_profile_api_v1_users_me_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserRead"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "users"
                ],
                "summary": "Update Own Profile",
                "operationId": "update_own_profile_api_v1_users_me_patch",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/v1/users/me/orders": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Read Own Orders",
                "operationId": "read_own_orders_api_v1_users_me_orders_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/OrderRead"
                                    },
                                    "type": "array",
                                    "title": "Response Read Own Orders Api V1 Users Me Orders Get"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/v1/users/me/payments": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Read Own Payments",
                "operationId": "read_own_payments_api_v1_users_me_payments_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/PaymentRead"
                                    },
                                    "type": "array",
                                    "title": "Response Read Own Payments Api V1 Users Me Payments Get"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/v1/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "List Users",
                "operationId": "list_users_api_v1_users_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserRead"
                                    },
                                    "title": "Response List Users Api V1 Users Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/{user_id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get User",
                "operationId": "get_user_api_v1_users__user_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "users"
                ],
                "summary": "Set User Active",
                "description": "Admin activate/deactivate a user. `is_active` is taken from the body;\n`full_name` is ignored here (that's the user's own via PATCH /users/me).",
                "operationId": "set_user_active_api_v1_users__user_id__patch",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "OrderItemRead": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "product_id": {
                        "type": "integer",
                        "title": "Product Id"
                    },
                    "quantity": {
                        "type": "integer",
                        "title": "Quantity"
                    },
                    "price": {
                        "type": "string",
                        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                        "title": "Price"
                    },
                    "subtotal": {
                        "type": "string",
                        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                        "title": "Subtotal"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "product_id",
                    "quantity",
                    "price",
                    "subtotal"
                ],
                "title": "OrderItemRead"
            },
            "OrderRead": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "user_id": {
                        "type": "integer",
                        "title": "User Id"
                    },
                    "total_amount": {
                        "type": "string",
                        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                        "title": "Total Amount"
                    },
                    "status": {
                        "$ref": "#/components/schemas/OrderStatus"
                    },
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/OrderItemRead"
                        },
                        "type": "array",
                        "title": "Items"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "user_id",
                    "total_amount",
                    "status",
                    "items",
                    "created_at"
                ],
                "title": "OrderRead"
            },
            "OrderStatus": {
                "type": "string",
                "enum": [
                    "pending",
                    "paid",
                    "canceled"
                ],
                "title": "OrderStatus"
            },
...
            "PaymentProvider": {
                "type": "string",
                "enum": [
                    "stripe",
                    "bkash"
                ],
                "title": "PaymentProvider"
            },
            "PaymentRead": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "order_id": {
                        "type": "integer",
                        "title": "Order Id"
                    },
                    "provider": {
                        "$ref": "#/components/schemas/PaymentProvider"
                    },
                    "transaction_id": {
                        "type": "string",
                        "title": "Transaction Id"
                    },
                    "status": {
                        "$ref": "#/components/schemas/PaymentStatus"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "order_id",
                    "provider",
                    "transaction_id",
                    "status",
                    "created_at"
                ],
                "title": "PaymentRead"
            },
            "PaymentStatus": {
                "type": "string",
                "enum": [
                    "pending",
                    "success",
                    "failed"
                ],
                "title": "PaymentStatus"
            },
...
            "UserRead": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "full_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Full Name"
                    },
                    "is_admin": {
                        "type": "boolean",
                        "title": "Is Admin"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "email",
                    "full_name",
                    "is_admin",
                    "is_active",
                    "created_at"
                ],
                "title": "UserRead"
            },
            "UserUpdate": {
                "properties": {
                    "full_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Full Name"
                    },
                    "is_active": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Is Active"
                    }
                },
                "type": "object",
                "title": "UserUpdate"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/v1/auth/login"
                    }
                }
            }
        }
    }
}